Deleted BaseRecipe and SelfCreatedRecipe.
