+++ src/entity/Recipe.java
@@ -2,20 +2,31 @@ package entity;
 
 import java.util.Collection;
 
-public class Recipe extends BaseRecipe{
+public class Recipe{
 
-    private final String url;
-    private final String imageUrl;
-    private final RecipeInfo info;
-    private final NutritionData nutrition;
+    private int ID;
+    private String title;
+    private String url;
+    private String imageUrl;
+    private RecipeInfo info;
+    private NutritionData nutrition;
     private Collection<Recipe> similarRecipes;
 
     public Recipe(int ID, String title, String url, String imageUrl, RecipeInfo info, NutritionData nutrition) {
-        super(ID, title);
+        this.ID = ID;
+        this.title = title;
         this.url = url;
         this.imageUrl = imageUrl;
         this.info = info;
         this.nutrition = nutrition;
+    }
+
+    public int getID() {
+        return ID;
+    }
+
+    public String getTitle() {
+        return title;
     }
 
     public String getUrl() {
@@ -40,6 +51,6 @@ public class Recipe extends BaseRecipe{
     //TODO Wright a toString method prepare for display.
     @Override
     public String toString(){
-        return super.getTitle() + "\n" + info.toString();
+        return title + "\n" + info.toString();
     }
 }
